requirements.txt
Pillow===8.4.0
fastapi===0.70.0
uvicorn===0.15.0
requests===2.26.0
python-multipart===0.0.5